Added iOS library target.

+++ src/kemarmic.cpp
@@ -95,7 +95,7 @@ Array<Array<std::vector<Sample>, MAX_NUM_AZIMUTHS>, NUM_ELEVATIONS>
       
       file.open (file_path, std::ios::in | std::ios::binary | std::ios::ate);
       if (! file.good()) { throw "Kemar lib not found."; }
-      long size = file.tellg();
+      long size = (long) file.tellg();
       assert(sizeof(short) == 2);
       short* data = new short[size/2];
